pyproject.toml
[project]
name = "marvin"
version = "0.1.0"
description = "42 discord general purpose bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.54.3",
    "pydantic-settings>=2.6.1",
]

[dependency-groups]
dev = [
    "pytest-xdist>=3.6.1",
]
